feat: add createEventWaitPromise
Creates createMaybeAbortablePromise which will also listen to a given
event.
Add its tests.

OBS: EventEmitter is relaxed and allows for multiple arguments, such as
`emitter.emit(1, 'abc')` will call your event handler with all of
those arguments.

In such case we should cooperate with that, otherwise we just get the
first argument and discard the rest