fix: battery and firmware request; remove requestCurrentConfig
In the BE, add proper Promise.resolve in requestFirmwareVersion and
requestBatteryLife for Vendor2Model21 and Vendor1 Model11 when in
wired mode.

Also in the BE, remove requestCurrentConfig method for all devices
and manager call.
Its possible results were not only not being listened by anyone,
but also because we are queueing the device calls in the
manager.service, calling requestCurrentConfig would request info
from the device without respecting the queue order.