feat: reconnect mouse on attach
Added the ability to detect mouse attach/detach and reload the
application. This allow to avoid asking the user to connect to
the mouse allowed previously.

In case if the device have serial ID or connected over BT it is possible
to attach/detach the device. But if the serial attribute is not
available only PWA reload without asking permission is working.

See rationale: https://bugs.chromium.org/p/chromium/issues/detail?id=958918

Signed-off-by: Denis Pynkin <denis.pynkin@collabora.com>